Create stream and utils package
