+++ contracts/streamswap-factory/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod contract;
+mod error;
 pub mod state;
