+++ contracts/streamswap-stream/src/lib.rs
@@ -5,8 +5,6 @@ pub mod contract;
 mod error;
 mod helpers;
 mod killswitch;
-pub mod msg;
 pub mod state;
 #[cfg(test)]
 mod tests;
-pub mod threshold;
