+++ packages/types/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod factory;
+pub mod stream;
