+++ testing/src/tests/streamswap_tests/resume_protocol_admin.rs
@@ -10,9 +10,9 @@ mod resume_protocol_admin {
     use cosmwasm_std::{coin, Addr, BlockInfo};
     use cw_multi_test::Executor;
     use streamswap_stream::state::Status;
-    use streamswap_stream::{
-        msg::{ExecuteMsg as StreamSwapExecuteMsg, QueryMsg as StreamSwapQueryMsg, StreamResponse},
-        ContractError as StreamSwapError,
+    use streamswap_stream::ContractError as StreamSwapError;
+    use streamswap_types::stream::msg::{
+        ExecuteMsg as StreamSwapExecuteMsg, QueryMsg as StreamSwapQueryMsg, StreamResponse,
     };
 
     #[test]
