+++ testing/src/tests/streamswap_tests/resume_stream.rs
@@ -10,9 +10,9 @@ mod resume_stream_test {
     use cosmwasm_std::{coin, Addr, BlockInfo};
     use cw_multi_test::Executor;
 
-    use streamswap_stream::{
-        msg::{ExecuteMsg as StreamSwapExecuteMsg, QueryMsg as StreamSwapQueryMsg, StreamResponse},
-        ContractError as StreamSwapError,
+    use streamswap_stream::ContractError as StreamSwapError;
+    use streamswap_types::stream::msg::{
+        ExecuteMsg as StreamSwapExecuteMsg, QueryMsg as StreamSwapQueryMsg, StreamResponse,
     };
 
     #[test]
