+++ testing/src/tests/streamswap_tests/update_stream.rs
@@ -12,10 +12,9 @@ mod update_stream_tests {
 
     use cosmwasm_std::{coin, Addr, BlockInfo, Decimal, Decimal256, Uint128};
     use cw_multi_test::Executor;
-    use streamswap_stream::{
-        msg::{ExecuteMsg as StreamSwapExecuteMsg, QueryMsg as StreamSwapQueryMsg, StreamResponse},
-        state::Status,
-        ContractError,
+    use streamswap_stream::{state::Status, ContractError};
+    use streamswap_types::stream::msg::{
+        ExecuteMsg as StreamSwapExecuteMsg, QueryMsg as StreamSwapQueryMsg, StreamResponse,
     };
 
     #[test]
@@ -165,7 +164,7 @@ mod update_stream_tests {
             .wrap()
             .query_wasm_smart(
                 Addr::unchecked(stream_swap_contract_address.clone()),
-                &streamswap_stream::msg::QueryMsg::Stream {},
+                &streamswap_types::stream::msg::QueryMsg::Stream {},
             )
             .unwrap();
 
@@ -193,7 +192,7 @@ mod update_stream_tests {
             .wrap()
             .query_wasm_smart(
                 Addr::unchecked(stream_swap_contract_address.clone()),
-                &streamswap_stream::msg::QueryMsg::Stream {},
+                &streamswap_types::stream::msg::QueryMsg::Stream {},
             )
             .unwrap();
